More Web UI layout changes
The labels in the tabs seemed a bit unnecessary, since they duplicate
the tab names. So I removed them.

The 'Refresh' and 'Export CSV' buttons seemed a bit misplaced at the
bottom, so I moved them up with the other buttons.

I added a 'Collapse' button to the jobs container.